enunciado02/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
